SparseArray templated on size_t for consistency that matters in Windows

+++ src/molpro/linalg/array/DistrArray.h
@@ -92,7 +92,7 @@ public:
   using distributed_array = void; //!< a compile time tag that this is a distributed array
   using value_type = double;
   using index_type = size_t;
-  using SparseArray = std::map<unsigned long int, double>;
+  using SparseArray = std::map<size_t, double>;
   using Distribution = util::Distribution<index_type>;
 
 protected:
